+++ src/constants/index.js
@@ -1,4 +1,4 @@
-const VERSION = '1.17.0'
+const VERSION = '1.17.1'
 
 let bootstrapVersion = 4
 try {
@@ -248,6 +248,7 @@ const DEFAULTS = {
   footerStyle (column) {
     return {}
   },
+  searchAccentNeutralise: false,
   showColumns: false,
   showColumnsToggleAll: false,
   showColumnsSearch: false,
@@ -291,7 +292,6 @@ const DEFAULTS = {
   buttonsPrefix: CONSTANTS.classes.buttonsPrefix,
   buttonsClass: CONSTANTS.classes.buttons,
   icons: CONSTANTS.icons,
-  html: CONSTANTS.html,
   iconSize: undefined,
   iconsPrefix: CONSTANTS.iconsPrefix, // glyphicon or fa(font-awesome)
   loadingFontSize: 'auto',